pyproject.toml
[project]
requires-python = ">=3.11"
name = "neotomaart"
description = "Article Recommendation system for Research Databases."
version = "0.0.1"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Simon Goring", email = "goring@wisc.edu" }
]
maintainers = [
    { name = "Simon Goring", email = "goring@wisc.edu" }
]
keywords = ["postgres", "sklearn", "embeddings", "recommender", "Docker"]
classifiers = ["Development Status :: 3 - Alpha", "Programming Language :: Python :: 3"]
dependencies = [
    "bs4>=0.0.2",
    "datetime>=5.5",
    "langdetect>=1.0.9",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.5.2",
]
[project.urls]
"Source code" = "https://github.com/NeotomaDB/article-relevance"
[tool]
[tool.hatch]
[tool.hatch.build.targets.wheel]
packages = ["src/article_relevance"]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
